LPS-168800 Rename getManualAddIconDataMap to getActionDropdownItems and return List<DropdownItem> instead of a map and then converting it to List<DropdownItem>
